+++ scalyr_agent/builtin_monitors/linux_system_metrics.py
@@ -21,7 +21,7 @@ import os
 import re
 import scalyr_agent.third_party.tcollector.tcollector as tcollector
 from Queue import Empty
-from scalyr_agent import ScalyrMonitor, BadMonitorConfiguration
+from scalyr_agent import ScalyrMonitor, BadMonitorConfiguration, define_metric, define_log_field
 from scalyr_agent.third_party.tcollector.tcollector import ReaderThread
 from scalyr_agent.json_lib import JsonObject
 from scalyr_agent import StoppableThread
@@ -29,131 +29,204 @@ from scalyr_agent import StoppableThread
 # In this file, we make use of the third party tcollector library.  Much of the machinery here is to
 # shoehorn the tcollector code into our framework.
 
+__monitor__ = __name__
+
 # These are the metrics collected, broken up by tcollector controllor:
 # We use the original tcollector documentation here has much as possible.
-#
-# From procstats:
-#   proc.stat.cpu type=*:        CPU counters in units of jiffies, where type can be one of user, nice, system, iowait,
-#                                irq, softirq, steal, guest.  As a rate, they should add up to 100*numcpus on the host.
-#   proc.stat.intr:              The number of interrupts since boot.
-#   proc.stat.ctxt               The number of context switches since boot.
-#   proc.stat.processes          The number of processes created since boot.
-#   proc.stat.procs_blocked      The number of processes currently blocked on I/O.
-#   proc.loadavg.1m              The load average over 1 minute.
-#   proc.loadavg.5m              The load average over 5 minutes.
-#   proc.loadavg.15m             The load average over 15 minutes.
-#   proc.loadavg.runnable        The number of runnable threads/processes.
-#   proc.loadavg.total_threads   The total number of threads/processes.
-#   proc.kernel.entropy_avail    The number of bits of entropy that can be read without blocking from /dev/random
-#   proc.uptime.total            The seconds since boot.
-#   proc.uptime.now              The seconds since boot of idle time
-#   proc.meminfo.*               Information about memory usage, in counts of 1 KB pages.  The notable categories are:
-#                                memtotal (the total amount of RAM), memfree (the total unused RAM),
-#                                buffers (the total amount of RAM used by system buffers), cached (the total amount
-#                                of RAM used for caching file blocks).  To calculate available memory, you should add
-#                                the memfree and cached values together.  See the following link for more details:
-#              http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/tree/Documentation/filesystems/proc.txt
-#                                for information about the other memory categories.
-#   proc.vmstat.pgfault          The number of minor page faults since boot.
-#   proc.vmstat.pgmajfault       The number of major page faults since boot
-#   proc.vmstat.pswpin           The number of processes swapped in since boot.
-#   proc.vmstat.pswpout          The number of processes swapped out since boot.
-#   proc.vmstat.pgppin           The number of pages swapped in since boot.
-#   proc.vmstat.pgpout           The number of pages swapped out in since boot.
-#
-#   sys.numa.zoneallocs type=*, node=*
-#                                Number of pages allocated from the preferred node, either type=hit or type=miss.
-#   sys.numa.foreign_allocs node=*
-#                                Number of pages allocated from node because the preferred node did not have any free.
-#   sys.numa.allocation node=* type=*
-#                                Number of pages allocated either type=locally or type=remotely for processes on this
-#                                node.
-#   sys.numa.interleave node=* type=hit
-#                                The number of pages allocated successfully by the interleave strategy.
-#
-# Metrics from netstat:
-#   net.sockstat.num_sockets     Number of sockets allocated (only TCP).
-#   net.sockstat.num_timewait    Number of TCP sockets currently in TIME_WAIT state.
-#   net.sockstat.sockets_inuse, type=*
-#                                Number of sockets in use by type.
-#   net.sockstat.num_orphans     Number of orphan TCP sockets (not attached to any file descriptor).
-#   net.sockstat.memory, type=*  Memory allocated for this socket type (in bytes).
-#   net.sockstat.ipfragqueues    Number of IP flows for which there are currently fragments queued for reassembly.
-#   net.stat.tcp.abort, type=*   Number of connections that the kernel had to abort due broken down by reason.
-#   net.stat.tcp.abort.failed    Number of times the kernel failed to abort a connection because it didn't even have
-#                                enough memory to reset it.
-#   net.stat.tcp.congestion.recovery, type=*
-#                                Number of times the kernel detected spurious retransmits and was able to recover part
-#                                or all of the CWND, broken down by how it recovered.
-#   net.stat.tcp.delayedack, type=*
-#                                Number of delayed ACKs sent of different types.
-#   net.stat.tcp.failed_accept, reason=*
-#                                Number of times a connection had to be dropped  after the 3WHS.  reason=full_acceptq
-#                                indicates that the application isn't accepting connections fast enough.  You should
-#                                see SYN cookies too.
-#   net.stat.tcp.invalid_sack, type=*
-#                                Number of invalid SACKs we saw of diff types. (requires Linux v2.6.24-rc1 or newer)
-#   net.stat.tcp.memory.pressure Number of times a socket entered the "memory pressure" mode.
-#   net.stat.tcp.memory.prune, type=*
-#                                Number of times a socket had to discard received data due to low memory conditions,
-#                                broken down by type.
-#   net.stat.tcp.packetloss.recovery, type=*
-#                                Number of times we recovered from packet loss by type of recovery (e.g. fast
-#                                retransmit vs SACK).
-#   net.stat.tcp.receive.queue.full
-#                                Number of times a received packet had to be dropped because the socket's receive
-#                                queue was full (requires Linux v2.6.34-rc2 or newer)
-#   net.stat.tcp.reording, detectedby=*
-#                                Number of times we detected re-ordering broken down by how.
-#   net.stat.tcp.syncookies, type=*
-#                                SYN cookies (both sent & received).
-#
-# Metrics from iostat:#
-#   iostat.disk.read_requests, dev=*
-#                                Number of reads completed by device
-#   iostat.disk.read_merged. dev=*
-#                                Number of reads merged by device
-#   iostat.disk.read_sectors, dev=*
-#                                Number of sectors read by device
-#   iostat.disk.msec_read, dev=*   Time in msec spent reading by device
-#   iostat.disk.write_requests, dev=*
-#                                Number of writes completed by device
-#   iostat.disk.write_merged, dev=*
-#                                Number of writes merged by device
-#   iostat.disk.write_sectors, dev=*
-#                                Number of sectors written by device
-#   iostat.disk.msec_write, dev=* Time in msec spent writing by device
-#   iostat.disk.ios_in_progress, dev=*
-#                                Number of I/O operations in progress by device
-#   iostat.disk.msec_total, dev=*  Time in msec doing I/O by device
-#   iostat.disk.msec_weighted_total, dev=*
-#                                Weighted time doing I/O (multiplied by ios_in_progress) by device
-#
-# Note, all of these stats are also published per partition under iostat.partition
-#
-# Metrics from dfstate:
-#   df.1kblocks.total, mount=*, fstype=*
-#                              The total size of the file system broken down by mount and filesystem type.
-#   df.1kblocks.used, mount=*, fstype=*
-#                              The number of blocks used broken down by mount and filesystem type.
-#   df.1kblocks.available, mount=*, fstype=*
-#                              The number of locks available broken down by mount and filesystem type.
-#   df.inodes.total, mount=*, fstype=*
-#                              The total number of inodes broken down by mount and filesystem type.
-#   df.inodes.used, mount=*, fstype=*
-#                              The number of used inodes broken down by mount and filesystem type.
-#   df.inodes.free, mount=*, fstype=*
-#                              The number of free inodes broken down by mount and filesystem type.
-#
-# Metrics from ifstat:
-#   proc.net.bytes, iface=*, direction=(in|out)
-#                              The number of bytes through the interface broken down by interface and direction.
-#   proc.net.packets, iface=*, direction=(in|out)
-#                              The number of packets through the interface broken down by interface and direction.
-#   proc.net.errs, iface=*, direction=(in|out)
-#                              The number of packet errors broken down by interface and direction.
-#   proc.net.dropped, iface=*, direction=(in|out)
-#                              The number of dropped backets broken down by interface and direction.
+define_metric(__monitor__, 'proc.stat.cpu',
+              'CPU counters in units of jiffies, where ``type`` can be one of ``user``, ``nice``, ``system``,  '
+              '``iowait``, ``irq``, ``softirq``, ``steal``, ``guest``.  As a rate, they should add up to  '
+              '``100*numcpus`` on the host.',
+              extra_fields={'type': ''}, cumulative=True, category='general')
+define_metric(__monitor__, 'proc.stat.intr:',
+              'The number of interrupts since boot.', cumulative=True, category='general')
+define_metric(__monitor__, 'proc.stat.ctxt',
+              'The number of context switches since boot.', cumulative=True, category='general')
+define_metric(__monitor__, 'proc.stat.processes',
+              'The number of processes created since boot.', cumulative=True, category='general')
+define_metric(__monitor__, 'proc.stat.procs_blocked',
+              'The number of processes currently blocked on I/O.', cumulative=True, category='general')
+define_metric(__monitor__, 'proc.loadavg.1m',
+              'The load average over 1 minute.', category='general')
+define_metric(__monitor__, 'proc.loadavg.5m',
+              'The load average over 5 minutes.', category='general')
+define_metric(__monitor__, 'proc.loadavg.15m',
+              'The load average over 15 minutes.', category='general')
+define_metric(__monitor__, 'proc.loadavg.runnable',
+              'The number of runnable threads/processes.', category='general')
+define_metric(__monitor__, 'proc.loadavg.total_threads',
+              'The total number of threads/processes.', category='general')
+define_metric(__monitor__, 'proc.kernel.entropy_avail',
+              'The number of bits of entropy that can be read without blocking from /dev/random', unit='bits',
+              category='general')
+define_metric(__monitor__, 'proc.uptime.total',
+              'The total number of seconds since boot.', cumulative=True, category='general')
+define_metric(__monitor__, 'proc.uptime.now',
+              'The seconds since boot of idle time', unit='seconds', cumulative=True, category='general')
+
+define_metric(__monitor__, 'proc.vmstat.pgfault',
+              'The total number of minor page faults since boot.', cumulative=True, category='virtual memory')
+define_metric(__monitor__, 'proc.vmstat.pgmajfault',
+              'The total number of major page faults since boot', cumulative=True, category='virtual memory')
+define_metric(__monitor__, 'proc.vmstat.pswpin',
+              'The total number of processes swapped in since boot.', cumulative=True, category='virtual memory')
+define_metric(__monitor__, 'proc.vmstat.pswpout',
+              'The total number of processes swapped out since boot.', cumulative=True, category='virtual memory')
+define_metric(__monitor__, 'proc.vmstat.pgppin',
+              'The total number of pages swapped in since boot.', cumulative=True, category='virtual memory')
+define_metric(__monitor__, 'proc.vmstat.pgpout',
+              'The total number of pages swapped out in since boot.', cumulative=True, category='virtual memory')
+
+define_metric(__monitor__, 'sys.numa.zoneallocs',
+              'The number of pages allocated from the preferred node, either type=hit or type=miss.',
+              extra_fields={'node': '', 'type': ''}, category='numa')
+define_metric(__monitor__, 'sys.numa.foreign_allocs',
+              'The number of pages allocated from node because the preferred node did not have any free.',
+              extra_fields={'node': ''}, category='numa')
+define_metric(__monitor__, 'sys.numa.allocation',
+              'The number of pages allocated either type=locally or type=remotely for processes on this node.',
+              extra_fields={'node': '', 'type': ''}, category='numa')
+define_metric(__monitor__, 'sys.numa.interleave',
+              'The number of pages allocated successfully by the interleave strategy.',
+              extra_fields={'node': '', 'type': 'hit'}, category='numa')
+
+
+define_metric(__monitor__, 'net.sockstat.num_sockets',
+              'The total number of sockets allocated (only TCP).', category='sockets')
+define_metric(__monitor__, 'net.sockstat.num_timewait',
+              'The total number of TCP sockets currently in TIME_WAIT state.', category='sockets')
+define_metric(__monitor__, 'net.sockstat.sockets_inuse',
+              'The total number of sockets in use by type.',
+              extra_fields={'type': ''}, category='sockets')
+define_metric(__monitor__, 'net.sockstat.num_orphans',
+              'The total number of orphan TCP sockets (not attached to any file descriptor).', category='sockets')
+define_metric(__monitor__, 'net.sockstat.memory',
+              'Memory allocated for this socket type (in bytes).',
+              extra_fields={'type': ''}, unit='bytes', category='sockets')
+define_metric(__monitor__, 'net.sockstat.ipfragqueues',
+              'The total number of IP flows for which there are currently fragments queued for reassembly.',
+              category='sockets')
+
+define_metric(__monitor__, 'net.stat.tcp.abort',
+              'The total number of connections that the kernel had to abort due broken down by reason.',
+              extra_fields={'type': ''}, cumulative=True, category='network')
+define_metric(__monitor__, 'net.stat.tcp.abort.failed',
+              'The total number of times the kernel failed to abort a connection because it didn\'t even have enough '
+              'memory to reset it.', cumulative=True, category='network')
+define_metric(__monitor__, 'net.stat.tcp.congestion.recovery',
+              'The number of times the kernel detected spurious retransmits and was able to recover part or all of the '
+              'CWND, broken down by how it recovered.',
+              extra_fields={'type': ''}, cumulative=True, category='network')
+define_metric(__monitor__, 'net.stat.tcp.delayedack',
+              'The number of delayed ACKs sent of different types.',
+              extra_fields={'type': ''}, cumulative=True, category='network')
+define_metric(__monitor__, 'net.stat.tcp.failed_accept',
+              'The number of times a connection had to be dropped  after the 3WHS.  reason=full_acceptq indicates that '
+              'the application isn\'t accepting connections fast enough.  You should see SYN cookies too.',
+              extra_fields={'reason': ''}, cumulative=True, category='network')
+define_metric(__monitor__, 'net.stat.tcp.invalid_sack',
+              'The number of invalid SACKs we saw of diff types. (requires Linux v2.6.24-rc1 or newer)',
+              extra_fields={'type': ''}, cumulative=True, category='network')
+define_metric(__monitor__, 'net.stat.tcp.memory.pressure',
+              'The number of times a socket entered the "memory pressure" mode.', cumulative=True, category='network')
+define_metric(__monitor__, 'net.stat.tcp.memory.prune',
+              'The number of times a socket had to discard received data due to low memory conditions, broken down by '
+              'type.',
+              extra_fields={'type': ''}, cumulative=True, category='network')
+define_metric(__monitor__, 'net.stat.tcp.packetloss.recovery',
+              'The number of times we recovered from packet loss by type of recovery (e.g. fast retransmit vs SACK).',
+              extra_fields={'type': ''}, cumulative=True, category='network')
+define_metric(__monitor__, 'net.stat.tcp.receive.queue.full',
+              'The number of times a received packet had to be dropped because the socket\'s receive queue was full '
+              '(requires Linux v2.6.34-rc2 or newer)', cumulative=True, category='network')
+define_metric(__monitor__, 'net.stat.tcp.reording',
+              'The number of times we detected re-ordering broken down by how.',
+              extra_fields={'detectedby': ''}, cumulative=True, category='network')
+define_metric(__monitor__, 'net.stat.tcp.syncookies',
+              'SYN cookies (both sent & received).',
+              extra_fields={'type': ''}, cumulative=True, category='network')
+
+define_metric(__monitor__, 'iostat.disk.read_requests',
+              'The total number of reads completed by device',
+              extra_fields={'dev': ''}, cumulative=True, category='disk requests')
+define_metric(__monitor__, 'iostat.disk.read_merged',
+              'The total number of reads merged by device',
+              extra_fields={'dev': ''}, cumulative=True, category='disk requests')
+define_metric(__monitor__, 'iostat.disk.read_sectors',
+              'The total number of sectors read by device',
+              extra_fields={'dev': ''}, cumulative=True, category='disk requests')
+define_metric(__monitor__, 'iostat.disk.msec_read',
+              'Time in msec spent reading by device',
+              extra_fields={'dev': ''}, unit='milliseconds', cumulative=True, category='disk requests')
+define_metric(__monitor__, 'iostat.disk.write_requests',
+              'The total number of writes completed by device',
+              extra_fields={'dev': ''}, cumulative=True, category='disk requests')
+define_metric(__monitor__, 'iostat.disk.write_merged',
+              'The total number of writes merged by device',
+              extra_fields={'dev': ''}, cumulative=True, category='disk requests')
+define_metric(__monitor__, 'iostat.disk.write_sectors',
+              'The total number of sectors written by device',
+              extra_fields={'dev': ''}, cumulative=True, category='disk requests')
+define_metric(__monitor__, 'iostat.disk.msec_write',
+              'The total time in milliseconds spent writing by device',
+              extra_fields={'dev': ''}, unit='milliseconds', cumulative=True, category='disk requests')
+define_metric(__monitor__, 'iostat.disk.ios_in_progress',
+              'The number of I/O operations in progress by device',
+              extra_fields={'dev': ''}, cumulative=True, category='disk requests')
+define_metric(__monitor__, 'iostat.disk.msec_total',
+              'The total time in milliseconds doing I/O by device.',
+              extra_fields={'dev': ''}, unit='milliseconds', cumulative=True, category='disk requests')
+define_metric(__monitor__, 'iostat.disk.msec_weighted_total',
+              'Weighted time doing I/O (multiplied by ios_in_progress) by device.',
+              extra_fields={'dev': ''}, unit='milliseconds', cumulative=True, category='disk requests')
+
+define_metric(__monitor__, 'df.1kblocks.total',
+              'The total size of the file system broken down by mount and filesystem type.',
+              extra_fields={'mount': '', 'fstype': ''}, unit='bytes:1024', category='disk resources')
+define_metric(__monitor__, 'df.1kblocks.used',
+              'The number of blocks used broken down by mount and filesystem type.',
+              extra_fields={'mount': '', 'fstype': ''}, unit='bytes:1024', category='disk resources')
+define_metric(__monitor__, 'df.1kblocks.available',
+              'The number of locks available broken down by mount and filesystem type.',
+              extra_fields={'mount': '', 'fstype': ''}, unit='bytes:1024', category='disk resources')
+define_metric(__monitor__, 'df.inodes.total',
+              'The number of inodes broken down by mount and filesystem type.',
+              extra_fields={'mount': '', 'fstype': ''}, category='disk resources')
+define_metric(__monitor__, 'df.inodes.used',
+              'The number of used inodes broken down by mount and filesystem type.',
+              extra_fields={'mount': '', 'fstype': ''}, category='disk resources')
+define_metric(__monitor__, 'df.inodes.free',
+              'The number of free inodes broken down by mount and filesystem type.',
+              extra_fields={'mount': '', 'fstype': ''}, category='disk resources')
+
+define_metric(__monitor__, 'proc.net.bytes',
+              'The total number of bytes transmitted through the interface broken down by interface and direction.',
+              extra_fields={'direction': '', 'iface': ''}, unit='bytes', cumulative=True, category='network interfaces')
+define_metric(__monitor__, 'proc.net.packets',
+              'The total number of packets transmitted through the interface broken down by interface and direction.',
+              extra_fields={'direction': '', 'iface': ''}, cumulative=True, category='network interfaces')
+define_metric(__monitor__, 'proc.net.errs',
+              'The total number of packet errors broken down by interface and direction.',
+              extra_fields={'direction': '', 'iface': ''}, cumulative=True, category='network interfaces')
+define_metric(__monitor__, 'proc.net.dropped',
+              'The total number of dropped packets broken down by interface and direction.',
+              extra_fields={'direction': '', 'iface': ''}, cumulative=True, category='network interfaces')
+
+define_metric(__monitor__, 'proc.meminfo.memtotal',
+              'The total number of 1 KB pages of RAM.', unit='bytes:1024', category='memory')
+define_metric(__monitor__, 'proc.meminfo.memfree',
+              'The total number of unused 1 KB pages of RAM. This does not include the number of cached pages which '
+              'can be used when allocating memory.', unit='bytes:1024', category='memory')
+define_metric(__monitor__, 'proc.meminfo.cached',
+              'The total number of 1 KB pages of RAM being used to cache blocks from the filesystem.  These can be '
+              'reclaimed as used to allocate memory as needed.', unit='bytes:1024', category='memory')
+define_metric(__monitor__, 'proc.meminfo.buffered',
+              'The total number of 1 KB pages of RAM being used in system buffers.', unit='bytes:1024',
+              category='memory')
+
+define_log_field(__monitor__, 'monitor', 'Always ``linux_system_metrics``.')
+define_log_field(__monitor__, 'metric', 'The name of a metric being measured, e.g. "proc.stat.cpu".')
+define_log_field(__monitor__, 'value', 'The metric value.')
 
 
 class TcollectorOptions(object):