+++ CHANGELOG.md
@@ -12,6 +12,9 @@ This release is under active development.  The release date is TBD and will be u
 New features:
 
 * Support for the MySQL monitor plugin
+* New ``print_monitor_doc.py`` tool for printing documentation for a monitor.  This tool is experimental.
+* New functions for defining and adding documentation for Scalyr Plugin Monitor configuration options, log entries,
+  and metrics.
 
 Bug fixes: